Moved composable functions into their owns files - Used a lazyColumn inside BundlePatchesDialog for better speeds which has done wonders
