+++ app/src/main/java/app/revanced/manager/ui/component/bundle/ImportBundleDialog.kt
@@ -1,4 +1,4 @@
-package app.revanced.manager.ui.component.sources
+package app.revanced.manager.ui.component.bundle
 
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -11,13 +11,10 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Topic
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -33,7 +30,6 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import app.revanced.manager.R
-import app.revanced.manager.ui.component.BundleTopBar
 import app.revanced.manager.util.APK_MIMETYPE
 import app.revanced.manager.util.JAR_MIMETYPE
 import app.revanced.manager.util.parseUrlOrNull
@@ -54,8 +50,8 @@ fun ImportBundleDialog(
     var patchBundle by rememberSaveable { mutableStateOf<Uri?>(null) }
     var integrations by rememberSaveable { mutableStateOf<Uri?>(null) }
 
-    val patchBundleText = if (patchBundle == null) "" else patchBundle.toString()
-    val integrationText = if (integrations == null) "" else integrations.toString()
+    val patchBundleText = patchBundle?.toString().orEmpty()
+    val integrationText = integrations?.toString().orEmpty()
 
     val inputsAreValid by remember {
         derivedStateOf {
@@ -195,82 +191,4 @@ fun BundleInfoListItem(
         },
         trailingContent = trailingContent,
     )
-}
-@Composable
-fun BundleTextContent(
-    name: String,
-    isImportPage: Boolean = true,
-    onNameChange: (String) -> Unit = {},
-    isLocal: Boolean,
-    remoteUrl: String,
-    onRemoteUrlChange: (String) -> Unit = {},
-    patchBundleText: String,
-    onPatchLauncherClick: () -> Unit = {},
-    integrationText: String = "",
-    onIntegrationLauncherClick: () -> Unit = {},
-) {
-    OutlinedTextField(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(bottom = 16.dp),
-        value = name,
-        onValueChange = onNameChange,
-        label = {
-            Text(stringResource(R.string.bundle_input_name))
-        }
-    )
-    if (!isLocal) {
-        OutlinedTextField(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 16.dp),
-            value = remoteUrl,
-            onValueChange = onRemoteUrlChange,
-            label = {
-                Text(stringResource(R.string.bundle_input_source_url))
-            }
-        )
-    } else {
-        if(isImportPage) {
-            OutlinedTextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 16.dp),
-                value = patchBundleText,
-                onValueChange = {},
-                label = {
-                    Text("Patches Source File")
-                },
-                trailingIcon = {
-                    IconButton(
-                        onClick = onPatchLauncherClick
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.Topic,
-                            contentDescription = null
-                        )
-                    }
-                }
-            )
-
-            OutlinedTextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = 16.dp),
-                value = integrationText,
-                onValueChange = {},
-                label = {
-                    Text("Integrations Source File")
-                },
-                trailingIcon = {
-                    IconButton(onClick = onIntegrationLauncherClick) {
-                        Icon(
-                            imageVector = Icons.Default.Topic,
-                            contentDescription = null
-                        )
-                    }
-                }
-            )
-        }
-    }
 }
