+++ app/src/main/res/values/strings.xml
@@ -90,11 +90,17 @@
     <string name="storage">Storage</string>
     <string name="tab_apps">Apps</string>
     <string name="tab_sources">Sources</string>
+    <string name="delete">Delete</string>
+    <string name="refresh">Refresh</string>
+    <string name="remote">Remote</string>
+    <string name="local">Local</string>
     <string name="reload_sources">Reload all sources</string>
     <string name="source_download_fail">Failed to download patch bundle: %s</string>
     <string name="source_replace_fail">Failed to load updated patch bundle: %s</string>
     <string name="source_replace_integrations_fail">Failed to update integrations: %s</string>
     <string name="no_patched_apps_found">No patched apps found</string>
+    <string name="No_patches_available_to_view">No patches available to view</string>
+    <string name="Patches_available_tap_to_view">%d Patches available, tap to view</string>
     <string name="unsupported_app">Unsupported app</string>
     <string name="unsupported_patches">Unsupported patches</string>
     <string name="universal_patches">Universal patches</string>
